% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{AlgoParams}
\alias{AlgoParams}
\title{Agorithm Parameter class}
\description{
A class that holds the parameters and upon assigning, the values of the
biological parameters the LBI algorithm requres
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AlgoParams-new}{\code{AlgoParams$new()}}
\item \href{#method-AlgoParams-clone}{\code{AlgoParams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlgoParams-new"></a>}}
\if{latex}{\out{\hypertarget{method-AlgoParams-new}{}}}
\subsection{Method \code{new()}}{
Initialise biological parameter class. Parameters are self-explanatory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlgoParams$new(binwidth = 1, linf = NULL, l50 = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlgoParams-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AlgoParams-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlgoParams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
