% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_lbi_utilities.R
\name{lb_ind_ices}
\alias{lb_ind_ices}
\title{Calculate the wanted LBI indicators}
\source{
https://raw.githubusercontent.com/ices-tools-dev/LBI_shiny/master/utilities.R
}
\usage{
lb_ind_ices(data, binwidth, linf, lmat, weight)
}
\arguments{
\item{data}{length-frequency matrix.}

\item{binwidth}{A number indicating the width of each bin.}

\item{linf}{A number indicating the asymptotic length.}

\item{lmat}{A number indicating the first sexual maturity.}

\item{weight}{weight-at-length matrix.}
}
\value{
A dataframe with all calculated indicators.
}
\description{
Generate a matrix with the calculated LBI indicators so that they can be compared with thresholds.
This is the untouched ICES function.
}
