% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_lbi_utilities.R
\name{lb_ind_pretty}
\alias{lb_ind_pretty}
\title{Calculate the wanted LBI indicators}
\usage{
lb_ind_pretty(data, binwidth, linf, lmat)
}
\arguments{
\item{data}{length-frequency matrix.}

\item{binwidth}{A number indicating the width of each bin.}

\item{linf}{A number indicating the asymptotic length.}

\item{lmat}{A number indicating the first sexual maturity.}
}
\value{
A dataframe with all calculated indicators
}
\description{
Generate a matrix with the calculated LBI indicators so that they can be compared with thresholds.
This is the modified version of 'lb_ind_ices' and does not allow to use the weight-at-length matrix needed only
to calculate the 'Optimal Yield' as the maximum length that holds the largest biomass. The reason
being is that there are two bugs as follows: a) the multiplication of the frequency by the weight-at-length matrices
is performed in such a way that not the same year columns are picked and multiplied in the two matrices - columns
are shifted - and b) multiple length classes may have the same biomass yielding a returned dimension error.
Given these bugs and the fact that the affected indicator - Lmaxy/Lopt- isn't ued in the the traffic-light
visualisation form -lb_table-, this indicator is no longer calculated in this version of the function.
}
