% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbi_algo.R
\name{LbiAlgo}
\alias{LbiAlgo}
\title{LBI Algorithm Class}
\value{
A dataframe with all calculated indicators.

A modified dataframe holding all calculated indicators.
}
\description{
class wrapping and calling the LBI algorithm
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LbiAlgo-new}{\code{LbiAlgo$new()}}
\item \href{#method-LbiAlgo-lbi}{\code{LbiAlgo$lbi()}}
\item \href{#method-LbiAlgo-run}{\code{LbiAlgo$run()}}
\item \href{#method-LbiAlgo-clone}{\code{LbiAlgo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbiAlgo-new"></a>}}
\if{latex}{\out{\hypertarget{method-LbiAlgo-new}{}}}
\subsection{Method \code{new()}}{
Initialise the LBI class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbiAlgo$new(data, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{length-frequency dataframe}

\item{\code{params}}{R6 object or named List with biological parameters and binwidth}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbiAlgo-lbi"></a>}}
\if{latex}{\out{\hypertarget{method-LbiAlgo-lbi}{}}}
\subsection{Method \code{lbi()}}{
Wrap up and call the actual LBI algorithm function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbiAlgo$lbi()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method can be called in isolation to get the exact dataframe that the original LBI function returns.
However, this method has been modified so that the weight-at-length matrix is never used and therefore
Lmaxy and Lmaxy/Lopt ratio is no longer available.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbiAlgo-run"></a>}}
\if{latex}{\out{\hypertarget{method-LbiAlgo-run}{}}}
\subsection{Method \code{run()}}{
Run LBI algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbiAlgo$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbiAlgo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LbiAlgo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbiAlgo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
