% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{OutputPlotter}
\alias{OutputPlotter}
\title{Class plotting LBI results}
\value{
a ggplot class

a ggplot class

A list of all built plots

A tower with all plots aligned
}
\description{
Results are plotted as a tower of plots, one on top of another
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OutputPlotter-new}{\code{OutputPlotter$new()}}
\item \href{#method-OutputPlotter-build_plots}{\code{OutputPlotter$build_plots()}}
\item \href{#method-OutputPlotter-clone}{\code{OutputPlotter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputPlotter-new"></a>}}
\if{latex}{\out{\hypertarget{method-OutputPlotter-new}{}}}
\subsection{Method \code{new()}}{
Initialise the OutputPlotter class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputPlotter$new(data, plot_context, thresholds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a dataframe with the LBI indicators}

\item{\code{plot_context}}{A R6 object or named List with parameteres that allow customising certain aspects of the plot}

\item{\code{thresholds}}{A name list with the thresholds against which the LBI indicators will be copmared}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputPlotter-build_plots"></a>}}
\if{latex}{\out{\hypertarget{method-OutputPlotter-build_plots}{}}}
\subsection{Method \code{build_plots()}}{
Build a tower of plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputPlotter$build_plots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputPlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OutputPlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
