% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{PlotContext}
\alias{PlotContext}
\title{Plot's context details class}
\description{
A class that holds all the aspects that shape the final plot visualising the
results from the LBI algorithm.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlotContext-new}{\code{PlotContext$new()}}
\item \href{#method-PlotContext-clone}{\code{PlotContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-PlotContext-new}{}}}
\subsection{Method \code{new()}}{
Initialise the PlotContext class. Default values for the input parameters are enough to graph the
results for the first time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotContext$new(
  main_title_size = 15,
  main_title = "LBI Scores",
  threshold_line_width = 1.2,
  x_title_size = 18,
  x_text_size = 12,
  x_text_angle = 90,
  y_title_size = 18,
  y_text_size = 12,
  y_text_angle = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlotContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
