% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{Thresholds}
\alias{Thresholds}
\title{Thresholds list}
\format{
An object of class \code{list} of length 6.
}
\usage{
Thresholds
}
\description{
A list that holds all thresholds over which the LBI indicators are measured against.
}
\keyword{datasets}
