% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{Thresholds}
\alias{Thresholds}
\title{Thresholds class}
\description{
A class that holds all the thresholds over which the LBI indicators are measured.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Thresholds-new}{\code{Thresholds$new()}}
\item \href{#method-Thresholds-clone}{\code{Thresholds$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Thresholds-new"></a>}}
\if{latex}{\out{\hypertarget{method-Thresholds-new}{}}}
\subsection{Method \code{new()}}{
Initialise the Threshold class. Default values for the input parameters are those provided by the
existing literature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Thresholds$new(
  Lc_Lmat = 1,
  L25_Lmat = 1,
  Lmax5_Linf = 0.8,
  Pmega = 0.3,
  Lmean_Lopt = 0.9,
  Lmean_Lfem = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Thresholds-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Thresholds-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Thresholds$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
